requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
notion-client==2.7.0
python-escpos==3.1
jinja2==3.1.6